+++ pyisy/__main__.py
@@ -40,10 +40,10 @@ async def main(url, username, password, tls_version, events, node_servers):
         # await isy.initialize(node_servers) TODO: revert
         await isy.initialize(
             nodes=False,
-            clock=True,
+            clock=False,
             programs=False,
-            variables=False,
-            networking=True,
+            variables=True,
+            networking=False,
             node_servers=False,
         )
     except (ISYInvalidAuthError, ISYConnectionError):
@@ -58,7 +58,9 @@ async def main(url, username, password, tls_version, events, node_servers):
         raise
 
     # Print a representation of all the Nodes
-    _LOGGER.debug(repr(isy.nodes))
+    # _LOGGER.debug(repr(isy.nodes))
+    _LOGGER.info(isy.networking["1"])
+    _LOGGER.info(repr(isy.networking["1"]))
     _LOGGER.info("Total Loading time: %.2fs", time.time() - t_0)
 
     node_changed_subscriber = None