+++ pyisy/helpers/__init__.py
@@ -3,7 +3,9 @@ from __future__ import annotations
 
 import datetime
 from random import getrandbits
+from typing import cast
 
+from pyisy.constants import ISY_VALUE_UNKNOWN, UOM_DOUBLE_TEMP, UOM_ISYV4_DEGREES
 from pyisy.helpers.xml import (
     attr_from_element,
     attr_from_xml,
@@ -16,6 +18,9 @@ __all__ = [
     "attr_from_xml",
     "attr_from_element",
     "value_from_nested_xml",
+    "now",
+    "random_uuid_hex",
+    "convert_isy_raw_value",
 ]
 
 
@@ -27,3 +32,29 @@ def now() -> datetime.datetime:
 def random_uuid_hex() -> str:
     """Generate a random UUID hex."""
     return "%032x" % getrandbits(32 * 4)
+
+
+def convert_isy_raw_value(
+    value: int | float | None,
+    uom: str | None,
+    precision: int | str,
+    fallback_precision: int | None = None,
+) -> float | int | None:
+    """Fix ISY Reported Values.
+
+    ISY provides float values as an integer and precision component.
+    Correct by shifting the decimal place left by the value of precision.
+    (e.g. value=2345, prec="2" == 23.45)
+
+    Insteon Thermostats report temperature in 0.5-deg precision as an int
+    by sending a value of 2 times the Temp. Correct by dividing by 2 here.
+    """
+    if value is None or value == ISY_VALUE_UNKNOWN:
+        return None
+    if uom in (UOM_DOUBLE_TEMP, UOM_ISYV4_DEGREES):
+        return round(float(value) / 2.0, 1)
+    if precision not in ("0", 0):
+        return cast(float, round(float(value) / 10 ** int(precision), int(precision)))
+    if fallback_precision:
+        return round(float(value), fallback_precision)
+    return value